+++ src/Resources/contao/languages/de/modules.php
@@ -10,3 +10,7 @@
  * Backend-Modul Übersetzungen
  */
 $GLOBALS['TL_LANG']['MOD']['fideid'] = array('FIDE-ID-Nummern', 'FIDE-Identifikationsnummern');
+/**
+ * Frontend-Module
+ */
+$GLOBALS['TL_LANG']['FMD']['fideid'] = array('FIDE-ID Antragsbestätigung', 'FIDE-ID Antragsbestätigung anzeigen');